requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
certifi==2026.2.25
click==8.3.1
fastapi==0.135.2
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
Mako==1.3.10
MarkupSafe==3.0.3
packaging==26.0
pluggy==1.6.0
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
pytest-asyncio>=0.23.0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==7.4.0
SQLAlchemy==2.0.48
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.42.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
